+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 ##
+- **CHANGED** For multi-source downsync/get the separator for paths is changed to | to avoid problems with path that contains spaces
+- **CHANGED** Made multi-path options separate from single-path
+  -- `--source-path` vs `--source-paths`
+  -- `--version-local-store-index-path` vs `--version-local-store-index-paths`
+- **FIXED** Network paths with spaces in them don't work - #252
+
+## v0.4.2
 - **FIXED** Networks share paths that starts with \\ no longer need to be manually escaped (fixes https://github.com/DanEngelbrecht/golongtail/issues/249)
 - **UPDATED** Update longtaillib to v0.4.2